requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
boto3==1.17.53
botocore==1.20.59
certifi==2020.12.5
channels-redis==3.2.0
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==3.0.2
defusedxml==0.7.1
django-allauth==0.44.0
docutils==0.17.1
easy-thumbnails==2.7.1
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
kombu==4.6.11
MarkupSafe==1.1.1
msgpack==1.0.2
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==8.2.0
pre-commit==2.12.1
psycopg2-binary==2.8.6
psycopg2==2.8.6
PyHamcrest==2.0.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests-oauthlib==1.3.0
requests==2.25.1
simplejson==3.17.2
tenant-schemas-celery==1.0.1
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.4
vine==1.3.0
whitenoise==5.2.0
zipp==3.4.1

s3transfer<0.4
six==1.15.0
Unipath==1.1
zope.interface==5.4.0
django-celery-results==1.1.2
django-cors-headers==3.7.0
django-filter==2.1.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-storages==1.7.1
django-tenants==2.2.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
Django==2.2.17
uvicorn[standard]==0.13.4
